Combine the Color and Pixel class into Pixel
